pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "trading-strands"
version = "0.0.1"
description = "Strategy-as-prompt trading agents on Strands + Bedrock AgentCore"
readme = "README.md"
license = { file = "LICENSE" }
requires-python = ">=3.12"
authors = [{ name = "Kelsea" }]
keywords = ["trading", "agents", "strands", "bedrock", "agentcore", "llm"]
classifiers = [
    "Development Status :: 2 - Pre-Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Office/Business :: Financial :: Investment",
]

dependencies = [
    # Agent framework
    "strands-agents>=0.5",
    "bedrock-agentcore>=0.1",
    # AWS
    "boto3>=1.35",
    # Brokers
    "robin-stocks>=3.4",
    "alpaca-py>=0.30",
    # Market data
    "yfinance>=0.2.50",
    # Core
    "pydantic>=2.9",
    "httpx>=0.27",
    "anyio>=4.6",
    "structlog>=24.4",
    "typer>=0.12",
    "rich>=13.9",
    "python-dotenv>=1.0",
    # Dashboard
    "fastapi>=0.115",
    "uvicorn>=0.32",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3",
    "pytest-asyncio>=0.24",
    "pytest-cov>=5.0",
    "ruff>=0.7",
    "mypy>=1.13",
    "types-requests",
]

[project.scripts]
trading-strands = "trading_strands.app:main"

[project.urls]
Documentation = "file://./docs/SPEC.md"

[tool.hatch.build.targets.wheel]
packages = ["src/trading_strands"]

[tool.ruff]
line-length = 100
target-version = "py312"
src = ["src", "tests"]

[tool.ruff.lint]
select = [
    "E", "F", "W",   # pycodestyle + pyflakes
    "I",              # isort
    "B",              # bugbear
    "UP",             # pyupgrade
    "SIM",            # simplify
    "RUF",            # ruff-specific
    "ASYNC",          # async lints
    "S",              # bandit (security)
]
ignore = [
    "S101",  # allow assert in tests
]

[tool.ruff.lint.per-file-ignores]
"tests/**/*.py" = ["S"]
"infra/**/*.py" = ["S"]

[tool.mypy]
python_version = "3.12"
strict = true
warn_unused_ignores = true
warn_redundant_casts = true
files = ["src"]

[[tool.mypy.overrides]]
module = [
    "robin_stocks.*",
    "yfinance.*",
    "uvicorn.*",
    "boto3.*",
    "botocore.*",
    "mypy_boto3_dynamodb.*",
]
ignore_missing_imports = true

[tool.pytest.ini_options]
minversion = "8.0"
testpaths = ["tests"]
asyncio_mode = "auto"
addopts = "-ra --strict-markers --strict-config"
